pyproject.toml
[project]
name = "OCR"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

################################################################################
# Poetry configuration
################################################################################

[tool.poetry]
name = "OCR"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
nltk = "^3.9.1"
lxml = "^5.3.1"
ketos = "^2.7.1"
kraken = "5.2.9"
pillow = "^11.1.0"
